Add application endpoint
